+++ example/lib/RunModelByCameraDemo.dart
@@ -1,0 +1,150 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:pytorch_lite/pigeon.dart';
+import 'package:pytorch_lite_example/ui/box_widget.dart';
+
+import 'ui/camera_view.dart';
+
+/// [RunModelByCameraDemo] stacks [CameraView] and [BoxWidget]s with bottom sheet for stats
+class RunModelByCameraDemo extends StatefulWidget {
+  @override
+  _RunModelByCameraDemoState createState() => _RunModelByCameraDemoState();
+}
+
+class _RunModelByCameraDemoState extends State<RunModelByCameraDemo> {
+  List<ResultObjectDetection?>? results;
+  String? classification;
+
+  /// Scaffold Key
+  GlobalKey<ScaffoldState> scaffoldKey = GlobalKey();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      key: scaffoldKey,
+      backgroundColor: Colors.black,
+      body: Stack(
+        children: <Widget>[
+          // Camera View
+          CameraView(resultsCallback, resultsCallbackClassification),
+
+          // Bounding boxes
+          boundingBoxes2(results),
+
+          // Heading
+          // Align(
+          //   alignment: Alignment.topLeft,
+          //   child: Container(
+          //     padding: EdgeInsets.only(top: 20),
+          //     child: Text(
+          //       'Object Detection Flutter',
+          //       textAlign: TextAlign.left,
+          //       style: TextStyle(
+          //         fontSize: 28,
+          //         fontWeight: FontWeight.bold,
+          //         color: Colors.deepOrangeAccent.withOpacity(0.6),
+          //       ),
+          //     ),
+          //   ),
+          // ),
+
+          //Bottom Sheet
+          Align(
+            alignment: Alignment.bottomCenter,
+            child: DraggableScrollableSheet(
+              initialChildSize: 0.4,
+              minChildSize: 0.1,
+              maxChildSize: 0.5,
+              builder: (_, ScrollController scrollController) => Container(
+                width: double.maxFinite,
+                decoration: BoxDecoration(
+                    color: Colors.white.withOpacity(0.9),
+                    borderRadius: BORDER_RADIUS_BOTTOM_SHEET),
+                child: SingleChildScrollView(
+                  controller: scrollController,
+                  child: Center(
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        Icon(Icons.keyboard_arrow_up,
+                            size: 48, color: Colors.orange),
+                        (classification != null)
+                            ? Padding(
+                                padding: const EdgeInsets.all(8.0),
+                                child: Column(
+                                  children: [
+                                    StatsRow(
+                                        'Classification:', '${classification}'),
+                                  ],
+                                ),
+                              )
+                            : Container()
+                      ],
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          )
+        ],
+      ),
+    );
+  }
+
+  /// Returns Stack of bounding boxes
+  Widget boundingBoxes2(List<ResultObjectDetection?>? results) {
+    if (results == null) {
+      return Container();
+    }
+    return Stack(
+      children: results.map((e) => BoxWidget(result: e!)).toList(),
+    );
+  }
+
+  void resultsCallback(List<ResultObjectDetection?> results) {
+    setState(() {
+      this.results = results;
+      results.forEach((element) {
+        print({
+          "rect": {
+            "left": element?.rect.left,
+            "top": element?.rect.top,
+            "width": element?.rect.width,
+            "height": element?.rect.height,
+            "right": element?.rect.right,
+            "bottom": element?.rect.bottom,
+          },
+        });
+      });
+    });
+  }
+
+  void resultsCallbackClassification(String classification) {
+    setState(() {
+      this.classification = classification;
+    });
+  }
+
+  static const BOTTOM_SHEET_RADIUS = Radius.circular(24.0);
+  static const BORDER_RADIUS_BOTTOM_SHEET = BorderRadius.only(
+      topLeft: BOTTOM_SHEET_RADIUS, topRight: BOTTOM_SHEET_RADIUS);
+}
+
+/// Row for one Stats field
+class StatsRow extends StatelessWidget {
+  final String left;
+  final String right;
+
+  StatsRow(this.left, this.right);
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 8.0),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [Text(left), Text(right)],
+      ),
+    );
+  }
+}